legacy/src/requirements.txt
flask==2.0.1
jinja2==3.0.3
werkzeug==2.0.3
markupsafe==2.0.1
itsdangerous==2.0.1
click==8.0.4
opencv-python-headless==4.5.3.56
numpy==1.19.5
RPi.GPIO==0.7.0
picamera==1.13
SpeechRecognition==3.8.1
pyaudio==0.2.11
pocketsphinx==0.1.15

# AI桌宠新增依赖
google-generativeai>=0.3.0
pvporcupine>=3.0.0
edge-tts>=6.1.0
pygame>=2.1.0
openai-whisper>=20231117
torch>=1.13.0
torchaudio>=0.13.0
jieba>=0.42.1 

# OLED显示系统依赖
adafruit-circuitpython-ssd1306>=2.12.0
adafruit-blinka>=8.0.0
Pillow>=9.0.0